requirements.txt
cffi==1.16.0
click==8.1.3
colorama==0.4.5
cryptography==41.0.4
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==10.1.0
protobuf==3.20.1
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==0.21.0
reportlab==4.0.6
Werkzeug==2.2.2
